pyproject.toml
[project]
name = "gmail-draft-loop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.21.0",
    "google-genai>=1.55.0",
]
